backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
httpx
python-jose[cryptography]
google-generativeai
python-dotenv
requests
pydantic
pydantic-settings
